# docs/openapi/adapter-v1.yaml
openapi: '3.0.3'
info:
  title: ExternalDNS RegRu Adapter API
  version: '1.0.0'
  description: >-
    Internal API for the ExternalDNS to Reg.ru adapter sidecar. Documents
    all endpoints exposed by cmd/sidecar/main.go including health, readiness,
    metrics, event intake, diagnostics, and admin controls.
servers:
  - url: http://localhost:8080
    description: Local sidecar (default port)
  - url: https://adapter.internal.example.com
    description: Internal adapter base URL (example)
paths:
...
  /adapter/v1/events:
    post:
      tags: [events]
      summary: Event intake
      description: >-
        Accepts ExternalDNS endpoint events, normalizes them into adapter
        operations, filters by zone-namespace mapping, applies FQDN templates,
        and enqueues for async processing by the worker pool.
      operationId: ingestEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DNSEndpointEvent'
      responses:
        '200':
          description: All events accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventIntakeResponse'
        '206':
          description: Partial success — some events accepted, some rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventIntakeResponse'
        '400':
          description: All events rejected (invalid JSON or all failed validation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventIntakeResponse'
        '405':
          description: Method not allowed (only POST)
...
  /adapter/v1/admin/safe-mode:
    post:
      tags: [admin]
      summary: Toggle safe-mode (no-write)
      description: >-
        Toggle safe-mode on the adapter. When enabled, write operations are
        disabled and queued for replay.
      operationId: setSafeMode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                  description: true to enable safe-mode (no write)
      responses:
        '200':
          description: Safe-mode updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '403':
          description: Forbidden (insufficient privileges)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Adapter-Token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    DNSEndpointEvent:
      type: object
      description: >-
        ExternalDNS endpoint event. Contains DNS record change information
        from Kubernetes resources.
      properties:
        kind:
          type: string
          example: DNSEndpoint
        namespace:
          type: string
          example: default
        name:
          type: string
          example: my-service
        dnsName:
          type: string
          example: app.example.com
        recordType:
          type: string
          enum: [A, AAAA, CNAME, TXT]
          example: A
        targets:
          type: array
          items:
            type: string
          example: ["10.0.0.1"]
        recordTTL:
          type: integer
          example: 300
        labels:
          type: object
          additionalProperties:
            type: string
    EventIntakeResponse:
      type: object
      properties:
        accepted:
          type: integer
          description: Number of events successfully enqueued
          example: 3
        errors:
          type: integer
          description: Number of events that failed validation
          example: 1
        error_details:
          type: array
          items:
            type: string
          description: Human-readable error messages for rejected events
      required: [accepted, errors]
...
    GenericSuccess:
      type: object
      properties:
        result:
          type: string
          example: success
    ErrorResponse:
      type: object
      properties:
        result:
          type: string
          example: error
        error_code:
          type: string
        error_text:
          type: string
security:
  - ApiKeyAuth: []
...